requirements.txt
environ
PyYAML
sqlalchemy
requests
bs4
pandas
numpy